requirements.txt
google-adk==1.15.1


openai